test/requirements.txt
cx_oracle
psycopg2-binary
pyodbc
